Add back in features and make options common
